refactor(feature-flags): Simplify implementation
