+++ src/System/Boot.tsx
@@ -10,8 +10,8 @@ import { CookieConsentManager } from "Components/CookieConsentManager/CookieCons
 import { PROGRESSIVE_ONBOARDING_KEYS } from "Components/ProgressiveOnboarding/progressiveOnboardingKeys"
 import { StickyProvider } from "Components/Sticky"
 import { ErrorBoundary } from "System/Components/ErrorBoundary"
-import { FeatureFlagProvider } from "System/Contexts/FeatureFlagContext"
 import { SystemContextProvider } from "System/Contexts/SystemContext"
+import { FeatureFlagProvider } from "System/FeatureFlags/FeatureFlagContext"
 import type { RouteProps } from "System/Router/Route"
 import type { ClientContext } from "System/Router/Utils/clientAppContext"
 import Events from "Utils/Events"
