+++ src/client.tsx
@@ -1,7 +1,7 @@
 import "Server/webpackPublicPath"
 import { loadableReady } from "@loadable/component"
 import { setupAnalytics } from "Server/analytics/helpers"
-import { getOrInitUnleashClient } from "Server/featureFlags/unleashClient"
+import { getOrInitUnleashClient } from "System/FeatureFlags/unleashClient"
 import { setupClientRouter } from "System/Router/clientRouter"
 import { setupSentryClient } from "System/Utils/setupSentryClient"
 import { setupWebVitals } from "System/Utils/setupWebVitals"
